Time: 0 ms (100%), Space: 7.8 MB (91.36%) - LeetHub

+++ 2579-count-total-number-of-colored-cells/2579-count-total-number-of-colored-cells.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long coloredCells(int n) {
-        int a=1;
+        long long a=1;
         while(n>1){
             a=a+4*(n-1);
             n--;
